diet_plan_builder_Full_stack/requirement.txt
uvicorn==0.34.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastavro==1.9.7
fastjsonschema==2.20.0
langchain==0.3.1
langchain-cohere==0.3.2
langchain-community==0.3.1
langchain-core==0.3.17
langchain-experimental==0.3.3
langchain-google-genai==2.0.7
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.20